Source code/server/requirements.txt
# FastAPI và ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# MQTT Client
paho-mqtt==1.6.1

# Serial Communication
pyserial==3.5

# HTTP Client cho OCR API
httpx==0.25.1
requests==2.31.0

# Database (SQLite)
sqlalchemy==2.0.23

# Template engine
jinja2==3.1.2

# Environment variables
python-dotenv==1.0.0

# Image processing
Pillow

# Async support
aiofiles==23.2.1
